Rename `mod` to `is-use-client`
